Not a recommended build, needs further testing, given the quantity of changes.
Lots of refactoring to solve issues with the design. Mostly to allow cleaner placement of the logic that determines where metadata is read from and written to when there are multiple sources.

Added new classes for Aperture, ShutterSpeed, ExposureBias, Rating and IsoSpeed. Placed the display logic for this types into their appropriate classes and removed it from the Provider classes.

Added support for LocationCreated (e.g. where the camera was when the photo was taken) vs LocationShown (e.g. what is shown in the photo). Added corrisponding Address & GpsPosition properties for these two concepts.

Added Interface to IFileMetadata to aid consistency

Added new methods to BitmapMetadataHelper to aid the detection of invalid query results.

Added new methods to FileMetadata to aid the reconciliation of different Metadata sources.

Added XmpPhotoshopProvider.

Updated and passed all Unit Tests.